refactor(verify): preserve readability thresholds
