+++ crates/temper-verify/src/cascade.rs
@@ -795,71 +795,6 @@ assert = "left_counter <= right_counter"
     }
 
     #[test]
-    fn capability_preflight_rejects_undeclared_counter_references() {
-        for assertion in ["missing > 0", "missing >= 0"] {
-            let spec = format!(
-                r#"
-[automaton]
-name = "UndeclaredCounterClaim"
-states = ["Active"]
-initial = "Active"
-allow_indefinite_states = ["Active"]
-
-[[invariant]]
-name = "CounterMustBeDeclared"
-when = ["Active"]
-assert = "{assertion}"
-"#
-            );
-
-            let result = VerificationCascade::from_ioa(&spec)
-                .with_sim_seeds(0)
-                .with_prop_test_cases(0)
-                .run();
-
-            assert!(!result.all_passed, "{assertion} must fail closed");
-            assert!(result.levels.is_empty());
-            assert_eq!(result.errors.len(), 1);
-            assert_eq!(result.errors[0].code, "TVE001");
-            assert_eq!(result.errors[0].assertion, assertion);
-        }
-    }
-
-    #[test]
-    fn capability_preflight_rejects_undeclared_status_references() {
-        for (when, assertion) in [("TypoState", "true"), ("Active", "never(TypoState)")] {
-            let spec = format!(
-                r#"
-[automaton]
-name = "UndeclaredStatusClaim"
-states = ["Active"]
-initial = "Active"
-allow_indefinite_states = ["Active"]
-
-[[invariant]]
-name = "StatusMustBeDeclared"
-when = ["{when}"]
-assert = "{assertion}"
-"#
-            );
-
-            let result = VerificationCascade::from_ioa(&spec)
-                .with_sim_seeds(0)
-                .with_prop_test_cases(0)
-                .run();
-
-            assert!(
-                !result.all_passed,
-                "when={when}, assert={assertion} must fail closed"
-            );
-            assert!(result.levels.is_empty());
-            assert_eq!(result.errors.len(), 1);
-            assert_eq!(result.errors[0].code, "TVE001");
-            assert_eq!(result.errors[0].assertion, assertion);
-        }
-    }
-
-    #[test]
     fn runtime_enforced_invariant_is_disclosed_without_claiming_model_proof() {
         let spec = r#"
 [automaton]
